pyproject.toml
[tool.poetry]
name = "bestemmissimo"
version = "0.1.0"
description = "Generate breathtaking blasphemies with this bot"
authors = ["Fabio Todaro <fbregist@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = {version = "^13.7", extras = ["ujson"]}
Pillow = "^8.3.2"
requests = "^2.26.0"
colorthief = "^0.2.1"
"discord.py" = "^1.7.3"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
flake8 = "^3.9.2"
mypy = "^0.910"
types-PyYAML = "^5.4.6"
types-requests = "^2.25.9"

[tool.poetry.scripts]
bestemmissimo = "bestemmissimo.cli.server:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
